changed delete game and added update game

+++ BLOCK223/src/ca/mcgill/ecse223/block/controller/Block223Controller.java
@@ -105,19 +105,38 @@ public class Block223Controller {
 	}
 
 	public static void deleteGame(String name) throws InvalidInputException {
+		UserRole currentUser = Block223Application.getCurrentUserRole();
+		Admin admin = currentGame.getAdmin();
 		if (!(currentUser instanceof Admin)) {
 			throw new InvalidInputException("Admin privileges are required to access game information.");
+		}
+		if (admin != (Admin) Block223Application.getCurrentUserRole()) {
+			throw new InavlidInputException("Only the admin who created the game can access its information.");
 		}
 		String error = "";
 		Game game = game.getWithName(name);
 		if (game != null) {
 			block223 = game.getBlock223();
 			game.delete();
-			BlockPersistence.save(block223);
+			Block223Persistence.save(block223);
 		}
 	}
 
 	public static void selectGame(String name) throws InvalidInputException {
+		String error = "";
+		UserRole currentUser = Block223Application.getCurrentUserRole();
+		Admin admin = currentGame.getAdmin();
+		if (!(currentUser instanceof Admin)) {
+			throw new InvalidInputException("Admin privileges are required to access game information.");
+		}
+		if (admin != (Admin) Block223Application.getCurrentUserRole()) {
+			throw new InavlidInputException("Only the admin who created the game can access its information.");
+		}
+		Game game = Game.getWithGame(name);
+		if (game == null) {
+			error = "A game with name" + name + "does not exist.";
+		}
+		BlockApplication.setCurrentGame(game);
 	}
 
 	public static void updateGame(String name, int nrLevels, int nrBlocksPerLevel, int minBallSpeedX, int minBallSpeedY,
@@ -162,11 +181,11 @@ public class Block223Controller {
 	// Query methods
 	// ****************************
 	public static List<TOGame> getDesignableGames() throws InvalidInputException {
-		if (!(currentUser instanceof Admin)) {
+		Block223 block223 = Block223Application.getBlock223();
+		Admin admin = Block223Application.getCurrentUserRole();
+		if (!(admin instanceof Admin)) {
 			throw new InvalidInputException("Admin privileges are required to access game information.");
 		}
-		Block223 block223 = BlockApplication.getBlock223();
-		Admin admin = BlockApplication.getCurrentUserRole();
 		List<TOGame> result = create();
 		Game games = getGames();
 		for (i=0; i<games.getSize(); i++) {
@@ -186,6 +205,27 @@ public class Block223Controller {
 	}
 
 	public static TOGame getCurrentDesignableGame() throws InvalidInputException {
+		UserRole currentUser = Block223Application.getCurrentUserRole();
+		Game currentgame = Block223Application.getCurrentGame();
+		Admin admin = currentGame.getAdmin();
+		if (!(currentUser instanceof Admin)) {
+			throw new InvalidInputException("Admin privileges are required to access game information.");
+		}
+		if (currentGame == null) {
+			throw new InvalidInputException("A game must be selected to access its information.");
+		}
+		if (admin != (Admin) Block223Application.getCurrentUserRole()) {
+			throw new InavlidInputException("Only the admin who created the game can access its information.");
+		}
+		Block223Application game = Block223Application.getCurrentGame();
+		TOGame to = create(game.getName(), game.getLevels().size(), 
+				game.getNrBlocksPerLevel(), 
+				game.getBall().getMinBallSpeedX(), 
+				game.getBall().getMinBallSpeedY(), 
+				game.getBall().getBallSpeedIncreaseFactor(), 
+				game.getPaddle().getMaxPaddleLength(), 
+				game.getPaddle().getMinPaddleLength());
+		return to;
 	}
 
 	public static List<TOBlock> getBlocksOfCurrentDesignableGame() throws InvalidInputException {
